SOCINT-266 Use /survey endpoint for available fulfilments Moved final RHSvc call
