+++ app/comms/__init__.py
@@ -1,4 +1,8 @@
+import re
 from app.request import RetryRequest
+from structlog import get_logger
+
+logger = get_logger('respondent-home')
 
 
 class MakeRequest:
@@ -21,3 +25,30 @@ class MakeRequest:
         """
         retry_request = RetryRequest(request, method, url, auth, headers, request_json, return_json)
         return await retry_request.make_request()
+
+
+class SingleClientIP:
+    @staticmethod
+    def single_client_ip(request):
+        if request['client_ip']:
+            client_ip = request['client_ip']
+            single_ip_validation_pattern = re.compile(r'^\d{1,3}\.\d{1,3}\.\d{1,3}\.\d{1,3}$')
+            if client_ip.count(',') > 1:
+                single_ip_value = client_ip.split(', ', -1)[-3]
+                if single_ip_validation_pattern.fullmatch(single_ip_value):
+                    single_ip = single_ip_value
+                else:
+                    logger.warn('clientIP failed validation. Provided IP - ' + client_ip,
+                                client_id=request['client_id'],
+                                trace=request['trace'])
+                    single_ip = ''
+            else:
+                logger.warn('clientIP failed validation. Provided IP - ' + client_ip,
+                            client_id=request['client_id'],
+                            trace=request['trace'])
+                single_ip = ''
+        elif request.headers.get('Origin', None) and 'localhost' in request.headers.get('Origin', None):
+            single_ip = '127.0.0.1'
+        else:
+            single_ip = ''
+        return single_ip
